requirements.txt
# To run the classifier

opencv-python
torch
numpy
torchvision
Pillow
facenet-pytorch

# To train the siamese model

pandas
seaborn
matplotlib